Clean up gulpfile.
