+++ lib/Core/ConsoleAnalytics.js
@@ -2,6 +2,7 @@
 
 /*global require*/
 var defined = require('terriajs-cesium/Source/Core/defined');
+//var foo = 'bar';
 
 var ConsoleAnalytics = function() {
     this.logToConsole = false;
